Attempt 1
This removes the new argument from the `indentline` method and instead
will map the highlight group of each column to the indentation level.
So with 5 levels of indent, you get `IndentLine1` ... `IndentLine5`.
There is a bit of a 'bug' / bad user experience though. You have to
define all highlights for all levels of indentation explicitly, i.e.
if you only define `IndentLine1`, `IndentLine2`, `IndentLine3`, then
only the first 3 indentation levels will have color, and the rest will
just be default.

So for this case, they will have to write some custom logic to map those
3 colors across however many indentation levels they want to cover,
which seems like a lot to put on consumers.