Attempt 2
For this attempt, I took a different approach of having a preset number
of indentation levels, I arbitrarily chose 6. This is similar to how
indent-blankline did things in a previous version. I don't like this
though because it restricts the consumer from additional customization,
and imposes a strict rule that you have to set all 6 or else it will use
the default mapping for ones you don't set.