Attempt 3
This time I just used neovim globals and existing highlight groups.
To use this one, you define the colors based on the pattern that
`indentmini` expcts. You can define as many or as few as you want.

The thing that I don't like about this is that it's a bit esoteric and
not declarative. But I think this does what @glepnir suggested as far as
having the consumer define the highlight groups.

Ex.

```lua
config = function()
  require('indentmini').setup()

  vim.cmd.highlight('IndentLine1 guifg=#AD7021')
  vim.cmd.highlight('IndentLine2 guifg=#8887C3')
  vim.cmd.highlight('IndentLine3 guifg=#738A05')
  vim.cmd.highlight('IndentLine4 guifg=#5F819D')
  vim.cmd.highlight('IndentLine5 guifg=#9E8FB2')
  vim.cmd.highlight('IndentLine6 guifg=#907AA9')
  vim.cmd.highlight('IndentLine7 guifg=#CDA869')
  vim.cmd.highlight('IndentLine8 guifg=#8F9D6A')
  -- And you could keep adding more
end
```